+++ doc/version_history.rst
@@ -1,10 +1,19 @@
-.. py:currentmodule:: lsst.ts.MTAlignment
+.. py:currentmodule:: lsst.ts.lasertracker
 
-.. _lsst.ts.MTAlignment.version_history:
+.. _lsst.ts.lasertracker.version_history:
 
 ###############
 Version History
 ###############
+
+v0.5.0
+------
+
+* Rename from ts_MTAlignment to ts_lasertracker and make an indexed component.
+  This requires ts_xml 15.
+
+  * Rename AlignmentCSC to LaserTrackerCsc.
+  * Rename AlignmentModel to T2SAModel.
 
 v0.4.1
 ------